requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.6
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
pillow==10.3.0
psycopg2==2.9.9
python-decouple==3.8
reportlab==4.2.0
sqlparse==0.5.0
tzdata==2024.1
